tests, names for urls, vacancy search fix
