+++ tracker/users/urls.py
@@ -7,14 +7,14 @@ from users.views import (
     FavoritesViewSet,
 )
 
-app_name = 'users'
+app_name = "users"
 
 router = routers.DefaultRouter()
-router.register(r"candidates", CandidateViewSet)
-router.register(r"favorites", FavoritesViewSet)
+router.register(r"candidates", CandidateViewSet, basename="candidates")
+router.register(r"favorites", FavoritesViewSet, basename="favorites")
 
 
 urlpatterns = [
-    path("candidates/<int:pk>/favorite/", APIAddFavorite.as_view()),
+    path("candidates/<int:pk>/favorite/", APIAddFavorite.as_view(), name="favorite"),
     path("", include(router.urls)),
 ]
