+++ tracker/vacancies/urls.py
@@ -3,7 +3,7 @@ from rest_framework import routers
 
 from vacancies.views import MatchCandidateViewSet, VacancyViewSet
 
-app_name = 'vacancies'
+app_name = "vacancies"
 
 router = routers.DefaultRouter()
 router.register(r"vacancies", VacancyViewSet)
@@ -11,6 +11,6 @@ router.register(r"vacancies", VacancyViewSet)
 urlpatterns = [
     path("", include(router.urls)),
     path(
-        "vacancies/<int:pk>/candidates/", MatchCandidateViewSet.as_view({"get": "list"})
+        "vacancies/<int:pk>/candidates/", MatchCandidateViewSet.as_view(), name="search"
     ),
 ]
